Timestamps that haven't been issued are set to 0

+++ Rush/GfxDeviceVK.h
@@ -425,7 +425,7 @@ public:
 
 		VkQueryPool      timestampPool = VK_NULL_HANDLE;
 		std::vector<u64> timestampPoolData;
-		std::vector<u8>  timestampSlotMap;
+		std::vector<u16> timestampSlotMap;
 		u32              timestampIssuedCount = 0;
 
 		DestructionQueue destructionQueue;
